==> 4_transcript_analysis.r <==
# -----------------------------------------
# Updated Date: 2014/03/24
# Input: The combined file generated by cbnCT.pl.
# Output: The statistical data calculated by NOISeq package.
# Environemt: Windows or Linux
# Description: Due to the in-complete analyses of fold change and expression difference, the NOISeq
#              added the normalization concept to analyze expression difference whether it is significant
#              or not as the global view.
# -----------------------------------------

# read data
getData <- read.table("3_combine.txt",header=T)
seqName <- getData[,1]
getData <- getData[2:3]
rownames(getData) <- seqName

# NOISeq analysis
source("http://bioconductor.org/biocLite.R")
biocLite("NOISeq")
library("NOISeq")

mfactors <- matrix(c("control","treatment"),nrow = 2, ncol = 1, byrow = TRUE, 
dimnames = list(c("control","treatment"),c("transcript")))

mydata <- readData(data=getData, factors=mfactors)
getNOIseqRes <- noiseq(mydata, k = 0.1, norm = "uqua", replicates = "no", factor="transcript", pnr = 0.2, nss = 10)

write.table(getNOIseqRes@results[[1]], file="4_transcript_ori.csv", sep=",", row.names=T, col.names=T, quote=F)
